tests/test_infrastructure_validation.py
"""Test infrastructure validation tests."""

import pytest
import sys
import os
from pathlib import Path


class TestInfrastructureValidation:
    """Validate that the testing infrastructure is properly set up."""
    
    def test_python_imports(self):
        """Test that basic Python imports work."""
        import json
        import numpy as np
        import yaml
        assert True
    
    def test_project_modules_importable(self):
        """Test that project modules can be imported."""
        # Add project root to path if not already there
        project_root = Path(__file__).parent.parent
        if str(project_root) not in sys.path:
            sys.path.insert(0, str(project_root))
        
        # Test importing project modules
        from T5.config import Config
        from utils.logger import init_logger
        assert True
    
    def test_fixtures_available(self, temp_dir, mock_config, mock_tokenizer):
        """Test that common fixtures are available and working."""
        assert os.path.exists(temp_dir)
        assert mock_config is not None
        assert mock_tokenizer is not None
        assert hasattr(mock_config, 'dataset')
        assert hasattr(mock_tokenizer, 'encode')
    
    def test_sample_data_fixtures(self, sample_dataset_item, sample_json_data):
        """Test that sample data fixtures work properly."""
        assert sample_dataset_item['question'] is not None
        assert sample_dataset_item['query'] is not None
        assert os.path.exists(sample_json_data)
    
    def test_database_fixtures(self, mock_sqlite_db, sample_tables_json):
        """Test that database fixtures work properly."""
        import sqlite3
        assert os.path.exists(mock_sqlite_db)
        assert os.path.exists(sample_tables_json)
        
        # Test database connectivity
        conn = sqlite3.connect(mock_sqlite_db)
        cursor = conn.cursor()
        cursor.execute("SELECT COUNT(*) FROM patients")
        result = cursor.fetchone()
        assert result[0] > 0
        conn.close()
    
    @pytest.mark.unit
    def test_unit_marker(self):
        """Test that unit test marker works."""
        assert True
    
    @pytest.mark.integration
    def test_integration_marker(self):
        """Test that integration test marker works."""
        assert True
    
    @pytest.mark.slow
    def test_slow_marker(self):
        """Test that slow test marker works."""
        assert True
    
    def test_temp_file_creation(self, temp_file):
        """Test that temporary file fixture works."""
        assert os.path.exists(temp_file)
        with open(temp_file, 'r') as f:
            content = f.read()
        assert content == "test content"
    
    def test_yaml_config_fixture(self, yaml_config_file):
        """Test that YAML config fixture works."""
        import yaml
        assert os.path.exists(yaml_config_file)
        
        with open(yaml_config_file, 'r') as f:
            config = yaml.safe_load(f)
        
        assert config['dataset'] == 'mimic_iii'
        assert config['model_name'] == 't5-base'
    
    def test_random_seed_fixture(self, set_random_seeds):
        """Test that random seed fixture works."""
        import torch
        import numpy as np
        
        # Test that seeds are set (no assertion needed, just verify no errors)
        torch.manual_seed(42)
        np.random.seed(42)
        assert True