pyproject.toml
[tool.poetry]
name = "ehrsql"
version = "0.1.0"
description = "EHRSQL: A Practical Text-to-SQL Benchmark for Electronic Health Records"
authors = ["EHRSQL Team <ehrsql@example.com>"]
readme = "README.md"
packages = [{include = "T5"}, {include = "utils"}, {include = "preprocess"}]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^2.0.0"
transformers = "^4.20.0"
numpy = "^1.21.0"
pyyaml = "^6.0"
tqdm = "^4.64.0"
func-timeout = "^4.3.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.0"

[tool.poetry.scripts]
test = "pytest:main"
tests = "pytest:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=T5",
    "--cov=utils", 
    "--cov=preprocess",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-report=xml:coverage.xml",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["T5", "utils", "preprocess"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/.*",
    "*/venv/*",
    "*/env/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"